python-telegram/requirements.txt
anyio==4.9.0
certifi==2025.4.26
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.1.0
python-telegram-bot==22.0
rsa==4.9.1
schedule==1.2.2
sniffio==1.3.1
Telethon==1.40.0
